Implement the feeds database table
