+++ app/Http/Controllers/DashboardController.php
@@ -10,26 +10,17 @@ class DashboardController extends Controller
 {
     use UserErrorTrait;
 
-    private FeedContract $feedDatabaseContract;
-
-    private UserContract $userContract;
-
     public function __construct(
-        FeedContract $feedDatabaseContract,
-        UserContract $userContract
+
     ) {
-        $this->feedDatabaseContract = $feedDatabaseContract;
-        $this->userContract = $userContract;
+
         $this->middleware('auth');
     }
 
 
     public function dashboard(): \Illuminate\View\View
     {
-        $feeds = $this->feedDatabaseContract->getAllFeeds(
-            $this->userContract
-        );
 
-        return view('dashboard', compact('feeds'));
+        return view('dashboard');
     }
 }
