+++ resources/js/bootstrap.js
@@ -3,6 +3,7 @@ window._ = require('lodash');
 try {
     require('bootstrap');
     require('datatables.net-bs5');
+    require('jquery');
 
 } catch (e) {}
