+++ resources/views/dashboard/articles.blade.php
@@ -22,6 +22,7 @@
                         {{$dataTable->table()}}
                     </div>
                 </div>
+
             </div>
 
         </div>
