+++ resources/views/partials/frontend/navbar.blade.php
@@ -3,6 +3,7 @@
         <a class="navbar-brand" href="{{ route('home') }}">
             {{ config('app.name', 'News Reader') }}
         </a>
+
         <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent"
                 aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation"><span
                 class="navbar-toggler-icon"></span></button>
@@ -22,6 +23,7 @@
                     @endif
                 @else
                     <li class="nav-item dropdown">
+
                         <a id="navbarDropdown" class="nav-link dropdown-toggle" href="#" role="button"
                            data-bs-toggle="dropdown" aria-haspopup="true" aria-expanded="false" v-pre>
                             {{ Auth::user()->name }}
